Major changes (events + controllers + config) for testing integration…

+++ database/migrations/2024_01_12_102723_create_parties_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('parties', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('video_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('video_id');
             $table->string('name');
             $table->datetime('start_time');
             $table->datetime('end_time')->nullable();
